+++ sources/packages/backend/apps/api/src/app.module.ts
@@ -50,7 +50,7 @@ import {
 import { LoggerModule } from "@sims/utilities/logger";
 import { ConfigModule } from "@sims/utilities/config";
 import { NotificationsModule } from "@sims/services/notifications";
-import { IER12IntegrationModule } from "@sims/integration/institution-integration/ier-integration/ier12-integration.module";
+import { IER12IntegrationModule } from "@sims/integrations/institution-integration/ier-integration/ier12-integration.module";
 
 @Module({
   imports: [
