+++ sources/packages/backend/apps/api/src/app.system-access.module.ts
@@ -65,8 +65,8 @@ import { FedRestrictionProcessingService } from "./esdc-integration/fed-restrict
 import { FedRestrictionIntegrationService } from "./esdc-integration/fed-restriction-integration/fed-restriction-integration.service";
 import { WorkflowClientService, SequenceControlService } from "@sims/services";
 // todo: once all integration are moved. remove sshservice.
-import { SshService } from "@sims/integration/services";
-import { IER12IntegrationModule } from "@sims/integration/institution-integration/ier-integration/ier12-integration.module";
+import { SshService } from "@sims/integrations/services";
+import { IER12IntegrationModule } from "@sims/integrations/institution-integration/ier-integration/ier12-integration.module";
 
 @Module({
   imports: [
