+++ sources/packages/backend/apps/api/src/cra-integration/__mocks__/SshServiceMock.ts
@@ -1,7 +1,7 @@
 import * as Client from "ssh2-sftp-client";
 import * as fs from "fs";
 import * as path from "path";
-import { SshService } from "@sims/integration/services";
+import { SshService } from "@sims/integrations/services";
 
 const MOCKED_RESPONSE_FILES = [
   "CCRA_RESPONSE_ABCSL00001.TXT",
