+++ sources/packages/backend/apps/api/src/cra-integration/cra-integration.module.ts
@@ -19,7 +19,7 @@ import {
 } from "../services";
 import { SequenceControlService, WorkflowClientService } from "@sims/services";
 import { ConfigModule } from "@sims/utilities/config";
-import { SshService } from "@sims/integration/services";
+import { SshService } from "@sims/integrations/services";
 
 @Module({
   imports: [AuthModule, ConfigModule],
