+++ sources/packages/backend/apps/api/src/esdc-integration/e-cert-integration/e-cert-full-time-integration/e-cert-full-time-integration.service.ts
@@ -19,8 +19,8 @@ import { DisbursementValueType, OfferingIntensity } from "@sims/sims-db";
 import { ECertIntegrationService } from "../e-cert-integration.service";
 import { ECertResponseRecord } from "../e-cert-files/e-cert-response-record";
 import { ConfigService } from "@sims/utilities/config";
-import { FixedFormatFileLine } from "@sims/integration/services/ssh";
-import { SshService } from "@sims/integration/services";
+import { FixedFormatFileLine } from "@sims/integrations/services/ssh";
+import { SshService } from "@sims/integrations/services";
 import { round } from "@sims/utilities";
 
 /**
