+++ sources/packages/backend/apps/api/src/esdc-integration/e-cert-integration/e-cert-integration.service.ts
@@ -6,8 +6,10 @@ import { ECertRecord } from "./models/e-cert-integration-model";
 import { OfferingIntensity } from "@sims/sims-db";
 import { ECertPartTimeResponseRecord } from "./e-cert-part-time-integration/e-cert-files/e-cert-response-record";
 import { ECertFullTimeResponseRecord } from "./e-cert-full-time-integration/e-cert-files/e-cert-response-record";
-import { SFTPIntegrationBase } from "@sims/integration/services";
-import { FixedFormatFileLine } from "@sims/integration/services/ssh";
+import {
+  FixedFormatFileLine,
+  SFTPIntegrationBase,
+} from "@sims/integrations/services/ssh";
 
 @Injectable()
 export abstract class ECertIntegrationService extends SFTPIntegrationBase<
