+++ sources/packages/backend/apps/api/src/esdc-integration/fed-restriction-integration/fed-restriction-integration.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { SshService } from "@sims/integration/services";
+import { SshService } from "@sims/integrations/services";
 import {
   FederalRestrictionService,
   RestrictionService,
