+++ sources/packages/backend/apps/api/src/esdc-integration/fed-restriction-integration/fed-restriction-integration.service.ts
@@ -1,5 +1,8 @@
 import { Injectable } from "@nestjs/common";
-import { SFTPIntegrationBase, SshService } from "@sims/integration/services";
+import {
+  SFTPIntegrationBase,
+  SshService,
+} from "@sims/integrations/services";
 import { ConfigService } from "@sims/utilities/config";
 import { FedRestrictionFileRecord } from "./fed-restriction-files/fed-restriction-file-record";
