+++ sources/packages/backend/apps/api/src/esdc-integration/sin-validation/sin-validation.module.ts
@@ -9,7 +9,7 @@ import { SequenceControlService } from "@sims/services";
 import { SINValidationIntegrationService } from "./sin-validation-integration.service";
 import { SINValidationProcessingService } from "./sin-validation-processing.service";
 import { ConfigModule } from "@sims/utilities/config";
-import { SshService } from "@sims/integration/services";
+import { SshService } from "@sims/integrations/services";
 
 @Module({
   imports: [AuthModule, ConfigModule],
