+++ sources/packages/backend/apps/api/src/sfas-integration/sfas-integration.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { SshService } from "@sims/integration/services";
+import { SshService } from "@sims/integrations/services";
 import { ConfigModule } from "@sims/utilities/config";
 import { AuthModule } from "../auth/auth.module";
 import {
