+++ sources/packages/backend/apps/api/test/jest-e2e.json
@@ -19,7 +19,7 @@
     "@sims/utilities": "<rootDir>/../../libs/utilities/src",
     "@sims/test-utils/(.*)": "<rootDir>/../../libs/test-utils/src/$1",
     "@sims/test-utils": "<rootDir>/../../libs/test-utils/src",
-    "@sims/integration/(.*)": "<rootDir>/../../libs/integration/src/$1",
-    "@sims/integration": "<rootDir>/../../libs/integration/src"
+    "@sims/integrations/(.*)": "<rootDir>/../../libs/integrations/src/$1",
+    "@sims/integrations": "<rootDir>/../../libs/integrations/src"
   }
 }
