+++ sources/packages/backend/libs/integrations/src/institution-integration/ier-integration/ier12-integration.module.ts
@@ -1,12 +1,12 @@
 import { Global, Module } from "@nestjs/common";
 import { WorkflowClientService } from "@sims/services";
-import {
-  SshService,
-  StudentAssessmentService,
-} from "@sims/integration/services";
 import { ConfigModule } from "@sims/utilities/config";
 import { IER12IntegrationService } from "./ier12-integration.service";
 import { IER12FileService } from "./ier12-file.service";
+import {
+  SshService,
+  StudentAssessmentService,
+} from "@sims/integrations/services";
 
 @Global()
 @Module({
