+++ sources/packages/backend/libs/integrations/src/institution-integration/ier-integration/ier12-integration.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { SFTPIntegrationBase, SshService } from "@sims/integration/services";
+import { SFTPIntegrationBase, SshService } from "@sims/integrations/services";
 import { ConfigService } from "@sims/utilities/config";
 import { IER12FileDetail } from "./ier12-file-detail";
 import { IER12FileLine, IER12Record } from "./models/ier12-integration.model";
